requirements.txt
skia-python
glfw
PyOpenGL
